requirements.txt
# list of all libraries we will need

transformers  #library to use hugging face 
transformers[sentencepiece]
datasets
sacrebleu   #library for row matrix(used for text summarization)
rouge_score
py7zr
pandas
nltk
tqdm 
PyYAML
matplotlib
torch #(if PC has GPU then install GPPU version )
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
-e .     #installs the file setup.py to insall the local package

